Rename DataAnalysisMockObjects file to MockObjects
