+++ qt/scientific_interfaces/Inelastic/test/Analysis/FqFitDataPresenterTest.h
@@ -14,9 +14,9 @@
 #include "Analysis/FqFitModel.h"
 #include "Analysis/FunctionBrowser/SingleFunctionTemplateBrowser.h"
 #include "Analysis/IndirectFitDataView.h"
-#include "DataAnalysisMockObjects.h"
 #include "IndirectAddWorkspaceDialog.h"
 #include "MantidFrameworkTestHelpers/IndirectFitDataCreationHelper.h"
+#include "MockObjects.h"
 
 using namespace Mantid::API;
 using namespace Mantid::IndirectFitDataCreationHelper;
